refactor: update navigation
