+++ src/app/app.module.ts
@@ -7,37 +7,43 @@ import { FormsModule }   from '@angular/forms';
 import { HttpClientModule }   from '@angular/common/http';
 
 import { AppComponent }   from './app.component';
+import { routing }   from './app.routing';
 
-import { NavComponent }   from './navigation/nav.component';
-import { FooterComponent}   from './footer/footer.component'
+import { NavComponent }   from './navigation';
+import { FooterComponent}   from './footer'
 
-import { HomeComponent }   from './home/home.component';
-import { NotFoundComponent} from './not-found/not-found.component';
+import { HomeComponent }   from './home';
+import { NotFoundComponent} from './not-found';
 
-import { SensorsComponent }   from './sensors/sensorts.component';
-import { RecordsComponent }   from './records/records.component';
-import { PatientsComponent }   from './patients/patients.component';
-import { ReportsComponent }   from './reports/reports.component';
+import { SensorsComponent }   from './sensors';
+import { RecordsComponent }   from './records';
+import { PatientsComponent }   from './patients';
+import { ReportsComponent }   from './reports';
 
-import { LoginComponent }   from './login/login.component';
-import { RegisterComponent }   from './register/register.component';
-
-
-const appRoutes: Routes =[
-    { path: 'home', component: HomeComponent},
-    { path: 'sensors', component: SensorsComponent},
-    { path: 'records', component: RecordsComponent},
-    { path: 'patients', component: PatientsComponent}, 
-    { path: 'reports', component: ReportsComponent},
-    { path: 'login', component: LoginComponent},
-    { path: 'register', component: RegisterComponent},
-    { path: '**', component: NotFoundComponent},
-];
+import { LoginComponent }   from './login';
+import { RegisterComponent }   from './register';
 
 @NgModule({
-    imports:      [ BrowserModule, FormsModule, HttpClientModule, RouterModule.forRoot(appRoutes)],
-    declarations: [ AppComponent, NavComponent, FooterComponent, HomeComponent, NotFoundComponent,SensorsComponent, 
-                    RecordsComponent, PatientsComponent, ReportsComponent, RegisterComponent, LoginComponent],
-    bootstrap:    [ AppComponent ]
+    imports: [
+        BrowserModule,
+        FormsModule,
+        HttpClientModule,
+        routing
+    ],
+    declarations: [
+        AppComponent,
+        NavComponent,
+        FooterComponent,
+        HomeComponent,
+        NotFoundComponent,
+        SensorsComponent, 
+        RecordsComponent,
+        PatientsComponent,
+        ReportsComponent,
+        RegisterComponent,
+        LoginComponent
+    ],
+    bootstrap: [ AppComponent ]
 })
+
 export class AppModule { }
