+++ src/app/footer/footer.component.html
@@ -1,5 +1,5 @@
 <footer class="page-footer font-small bg-light">
     <div class="footer-copyright text-center py-3">© 2020 Copyright:
-        <a href="https://github.com/aslamovyura/iCare-single-page-app/blob/master/LICENSE"> iCare</a>
+        <a href="https://github.com/aslamovyura/iCare-single-page-app/blob/master"> iCare</a>
     </div>
 </footer>
