+++ src/app/not-found/index.ts
@@ -1,0 +1,1 @@
+export * from './not-found.component';
