+++ src/app/patients/index.ts
@@ -1,0 +1,1 @@
+export * from './patients.component';
