+++ src/app/records/index.ts
@@ -1,0 +1,1 @@
+export * from './records.component';
