+++ src/app/register/index.ts
@@ -1,0 +1,1 @@
+export * from './register.component';
