+++ src/app/reports/index.ts
@@ -1,0 +1,1 @@
+export * from './reports.component';
