+++ src/app/sensors/index.ts
@@ -1,0 +1,1 @@
+export * from './sensors.component';
