+++ src/app/sensors/sensors.component.ts
